[meta-f] store values as tagged pointers

+++ meta-rs/meta-f/src/cps.rs
@@ -26,7 +26,7 @@ impl VarGen {
 pub(crate) enum Value {
     Var(Var),
     Label(Var),
-    Int(i64),
+    Int(i32),
     String(String),
 }
